pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ansys-modelcenter-workflow"
version = "0.1.dev0"

description = "Python API for using ModelCenter Desktop"
readme = "README.rst"
homepage = "https://github.com/pyansys/modelcenter/"

authors = ["Ansys, Inc."] # this is required
maintainers = ["PyAnsys developers <pyansys.support@ansys.com>"] # You can change the name. The email group works.
license = "MIT"

# Less than critical but helpful
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows :: Windows 7",
    "Operating System :: Microsoft :: Windows :: Windows Vista",
    "Operating System :: Microsoft :: Windows :: Windows 8",
    "Operating System :: Microsoft :: Windows :: Windows 8.1",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: Microsoft :: Windows :: Windows Server 2003",
    "Operating System :: Microsoft :: Windows :: Windows Server 2008",
]

packages = [
    { include = "ansys" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.22.3"
overrides = "^6.1"

# This is the closest version to what we're using in C++.
# It's probably not strictly necessary to keep these exactly in sync,
# but it might be helpful.
grpcio = "1.51.3"
grpcio-tools = "1.51.3"

Sphinx =  {version="^4.4", optional=true}
numpydoc =  {version="^1.2", optional=true}
ansys-sphinx-theme =  {version="^0.9", optional=true}
sphinx-copybutton =  {version="^0.5", optional=true}

codespell = {version="^2.1", optional=true}
flake8 = {version="^3.9", optional=true}

pytest = {version="^7.0", optional=true}
pytest-cov = {version="^3.0", optional=true}
pre-commit = {version="^2.17.0", optional=true}
mypy = {version="^0.940", optional=true}

ansys-engineeringworkflow-api = {git = "https://github.com/pyansys/ansys-engineeringworkflow-api.git", rev = "e66d11bbfa1175130913c7a7248db69ce002b341"}
ansys-common-variableinterop = {git = "https://github.com/pyansys/ansys-common-variableinterop.git", rev = "3c0a8f7"}

[tool.poetry.extras]
dev = ["Sphinx", "numpydoc", "ansys-sphinx-theme", "sphinx-copybutton", "codespell",
    "flake8", "pytest", "pytest-cov", "pre-commit", "mypy"]
docs = ["Sphinx", "numpydoc", "ansys-sphinx-theme", "sphinx-copybutton"]
style = ["codespell", "flake8"]
test = ["pytest", "pytest-cov"]
